Rename human input hook to orchestration request info
